More sounds
Changed the main blaster to medblaster so I could give it a separate sound from the enemy blaster.  

I've been adding sounds by finding the desired prefab and creating a new audio component, choosing the sound and routing it through the SFX mixer.  This gives me volume control of each individual sound, which allows me to balance it against the rest of the sounds.

Added impact sounds to medblaster, medblaster2 and drunkenrocket2.  Could not raise volume of impact sound, please show me how to do that.

Added Mod Acquired 1 sound to all mods on awake except default mod.

Added shield up a1 and b1 to Shield and ReflectorShield respectively

Possible solution to the volume jump problem:  Only have one of any type of sound be able to play at once. 

IE if you get the tri-blast upgrade, every time you fire it, the sound plays for only one of the blasters.  Multiple mods of the same type of weapon would work the same, only hearing one of them at the moment of firing.

I think this would work well, since the only thing that happens when multiple of the same sound plays is a volume increase.